cleanup some mistakes from ai
